+++ Lib/ToolsDir/loader.ts
@@ -56,10 +56,11 @@ R.r = function (name: string, parent: string) {
     var lp = p.toLowerCase();
     var m = R.m.get(lp);
     if (m == null && /\.js$/.test(lp)) {
-        m = R.m.get(lp.substr(0, lp.length - 3));
+        m = R.m.get(lp.slice(0, lp.length - 3));
     }
     if (m == null) {
         m = R.m.get(lp + "/index");
+        if (m != null) p = p + "/index";
     }
     if (m == null) throw new Error("Module " + name + " in " + (parent || "/") + " not registered");
     if (m.exports !== undefined) return m.exports;